# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Analytics","description":"Endpoints for viewing URL analytics."},{"name":"User Authentication","description":"Endpoints for user registration and login."},{"name":"User Profile Management","description":"Endpoints for users to manage their own profiles."},{"name":"URL Management","description":"Endpoints for url Submit a short URL and receive expanded details. Must cache results in Redis."}],"paths":{"/api/users/edit":{"put":{"tags":["User Profile Management"],"summary":"Update a user's profile","description":"Allows an authenticated user to update their own profile information (e.g., email, password).","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"200":{"description":"Profile updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}},"400":{"description":"Invalid data provided","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}},"401":{"description":"User is not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/urls/v1/expand":{"post":{"tags":["URL Management"],"summary":"Submit a short URL and receive expanded details","operationId":"expand","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UrlRequestDTO"}}},"required":true},"responses":{"200":{"description":"Receive expanded details successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UrlRespondDTO"}}}},"400":{"description":"Invalid data provided","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UrlRespondDTO"}}}},"401":{"description":"User is not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UrlRespondDTO"}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/auth/register":{"post":{"tags":["User Authentication"],"summary":"Register a new user","description":"Creates a new user account and an associated empty portfolio. Returns the created user's details.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"201":{"description":"User successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}},"400":{"description":"Invalid user data provided or username is already taken","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}}}}},"/api/auth/login":{"post":{"tags":["User Authentication"],"summary":"Authenticate a user","description":"Verifies a user's credentials and returns a JWT (JSON Web Token) upon success.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"200":{"description":"Authentication successful, JWT is returned","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid credentials provided","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/analytics/trending":{"get":{"tags":["Analytics"],"summary":"Fetch the top 10 expanded details from the Submitted short URLs sorted set","operationId":"trending","responses":{"200":{"description":"Receive the top 10 successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DomainTrendDTO"}}}}},"400":{"description":"Invalid data provided","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DomainTrendDTO"}}}}},"401":{"description":"User is not authenticated","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DomainTrendDTO"}}}}}},"security":[{"Bearer Authentication":[]}]}},"/api/users/me":{"get":{"tags":["User Profile Management"],"summary":"Get current user's details","description":"Fetches the profile information for the currently authenticated user based on their JWT.","operationId":"getUser","responses":{"200":{"description":"User details retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}},"401":{"description":"User is not authenticated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserRespondDTO"}}}}},"security":[{"Bearer Authentication":[]}]}}},"components":{"schemas":{"UserRequestDTO":{"type":"object","properties":{"username":{"type":"string","description":"Username for the account.","example":"modhtom","maxLength":20,"minLength":3},"mail":{"type":"string","description":"User's email address.","example":"modhtom@example.com"},"password_unHashed":{"type":"string","description":"User's password (plain text).","example":"password123","maxLength":2147483647,"minLength":6}},"required":["mail","password_unHashed","username"]},"UserRespondDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"Unique identifier for the user.","example":"a1b2c3d4-e5f6-7890-1234-567890abcdef"},"username":{"type":"string","description":"Username of the user.","example":"modhtom"},"mail":{"type":"string","description":"Email address of the user.","example":"modhtom@example.com"},"enabled":{"type":"boolean","description":"Indicates if the user's account is active.","example":true},"created_at":{"type":"string","format":"date-time","description":"Timestamp of when the user was created."},"updated_at":{"type":"string","format":"date-time","description":"Timestamp of the last update to the user's profile."}}},"UrlRequestDTO":{"type":"object","properties":{"url":{"type":"string","description":"The url to get expand.","example":"https://tinyurl.com/34775fb4"}},"required":["url"]},"UrlRespondDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"An ID for the url that got expand.","example":"modhtom.com"},"shortUrl":{"type":"string","description":"The URL that needs to be expand.","example":"modhtom.com"},"finalUrl":{"type":"string","description":"The URL that got expand.","example":"modhtom.com"},"domain":{"type":"string","description":" Domain Name for the url that got expand.","example":"modhtom.com"},"title":{"type":"string","description":"Domain Title for the url that got expand.","example":"modhtom.com"},"description":{"type":"string","description":"Domain Description for the url that got expand.","example":"modhtom.com"}},"required":["description","domain","finalUrl","id","shortUrl","title"]},"DomainTrendDTO":{"type":"object","properties":{"domain":{"type":"string","description":"Trending Domain Name.","example":"modhtom.com"},"count":{"type":"number","format":"double","description":"Trending Domain Count.","example":1}},"required":["count","domain"]}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}